% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.check.R
\name{quiz.check}
\alias{quiz.check}
\title{Check a quiz answer}
\usage{
quiz.check(ans)
}
\arguments{
\item{ans}{An answer. E.g., "A", or 3.14}
}
\value{
No value is returned, but message is displayed.
}
\description{
Check a quiz answer
}
\examples{
.tot = 0
.correct.answer1.1 = "A"
.hint1.1 = "'2+2'"
.explanation1.1 = "Arithmetic"
## Q1.1 What is 2 + 2?
## A. 4
## B. 3
## Replace the NA with your answer (e.g., 'A' in quotes)"
answer1.1 = NA
quiz.check(answer1.1)
answer1.1 = "A"
quiz.check(answer1.1)
answer1.1 = "B"
quiz.check(answer1.1)
answer1.1 = "C"
quiz.check(answer1.1)
answer1.1 = 4
quiz.check(answer1.1)
}
