% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizify.R
\name{quizify}
\alias{quizify}
\title{Create self-graded quizes within .Rmd files}
\usage{
quizify(initial.filename, final.filename, answer.key.filename)
}
\arguments{
\item{initial.filename}{Name of markdown file with quiz questions in {{...}} markup. E.g., quiz_initial.Rmd}

\item{final.filename}{Name of markdown file to output to. E.g., quiz_final.Rmd}

\item{answer.key.filename}{An intermediate file (will be deprecated someday)}
}
\value{
No value is returned, but final file is created.
}
\description{
Create self-graded quizes within .Rmd files
}
\examples{
initial.filename = system.file("arithmetic_quiz_initial.Rmd",
                               package = "quizify")
quizify(initial.filename,
        final.filename = "arithmetic_quiz_final.Rmd",
        answer.key.filename = "check_arithmetic_quiz.R")
# can cat to screen
x = scan("arithmetic_quiz_final.Rmd", character(), sep = "\\n")
cat(x, sep = "\\n")
file.remove("arithmetic_quiz_final.Rmd")
}
